[add] fix sonar smell bugs,and move DescriptionHolder to holder folder.
