+++ src/main/java/com/koy/kbot/beanprocessor/DescriptionBeanProcessor.java
@@ -1,10 +1,10 @@
 package com.koy.kbot.beanprocessor;
 
 import com.koy.kbot.configuration.core.Description;
-import com.koy.kbot.configuration.core.DescriptionHolder;
+import com.koy.kbot.holder.DescriptionHolder;
 import com.koy.kbot.configuration.core.Plugin;
 import com.koy.kbot.plugins.IPlugin;
-import org.springframework.beans.BeansException;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.stereotype.Component;
 
@@ -17,16 +17,19 @@ import org.springframework.stereotype.Component;
 @Component
 public class DescriptionBeanProcessor implements BeanPostProcessor {
 
+    @Autowired
+    private DescriptionHolder descriptionHolder;
+
     /**
      * before bean initialization,put plugin info into map
      */
     @Override
-    public Object postProcessBeforeInitialization(Object bean, String beanName) throws BeansException {
+    public Object postProcessBeforeInitialization(Object bean, String beanName){
         if(bean instanceof IPlugin){
             Description description = bean.getClass().getAnnotation(Description.class);
             Plugin plugin = bean.getClass().getAnnotation(Plugin.class);
             if(null != description && null != plugin){
-                DescriptionHolder.putDescription(description,plugin);
+                descriptionHolder.putDescription(description,plugin);
             }
         }
         return bean;
