+++ src/main/java/com/koy/kbot/holder/DescriptionHolder.java
@@ -1,8 +1,9 @@
-package com.koy.kbot.configuration.core;
+package com.koy.kbot.holder;
 
+import com.koy.kbot.configuration.core.Description;
+import com.koy.kbot.configuration.core.Plugin;
 import com.koy.kbot.entity.description.DescriptionInfo;
 import org.jetbrains.annotations.NotNull;
-import org.springframework.stereotype.Component;
 
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
@@ -14,17 +15,16 @@ import java.util.concurrent.ConcurrentMap;
  *
  * @author sivyer9303
  */
-@Component
 public class DescriptionHolder {
     /**
      * description info map
      */
-    private static final ConcurrentMap<String, DescriptionInfo> DESCRIPTION_MAP = new ConcurrentHashMap(16);
+    private final ConcurrentMap<String, DescriptionInfo> DESCRIPTION_MAP = new ConcurrentHashMap<>(16);
 
     /**
      * put description into map
      */
-    public static void putDescription(Description description,Plugin plugin){
+    public void putDescription(Description description, Plugin plugin){
         DescriptionInfo info = formatAnnotations(description);
         DESCRIPTION_MAP.putIfAbsent(plugin.name(), info);
     }
@@ -32,18 +32,16 @@ public class DescriptionHolder {
     /**
      * get descriptions
      */
-    public static String getDescriptions(){
+    public String getDescriptions(){
         StringBuilder stringBuilder = new StringBuilder();
-        DESCRIPTION_MAP.forEach((name,info) -> {
-            stringBuilder.append(name)
-                    .append(":")
-                    .append(info.getDesc())
-                    .append(",")
-                    .append("example usage is '")
-                    .append(info.getExample())
-                    .append("'")
-                    .append("\n");
-        });
+        DESCRIPTION_MAP.forEach((name,info) -> stringBuilder.append(name)
+                .append(":")
+                .append(info.getDesc())
+                .append(",")
+                .append("example usage is '")
+                .append(info.getExample())
+                .append("'")
+                .append("\n"));
         return stringBuilder.toString();
     }
 
@@ -52,7 +50,7 @@ public class DescriptionHolder {
      * @param description see {@link Description}
      * @return DescriptionInfo
      */
-    private static DescriptionInfo formatAnnotations(@NotNull Description description) {
+    private DescriptionInfo formatAnnotations(@NotNull Description description) {
         return DescriptionInfo.builder()
                 .author(description.author())
                 .desc(description.desc())
